// interfaces/registration-interface.ts
export default interface typeRegistrationProps {
  salutation: string;
  name: string;
  email: string;
  password: string;
  contact_no: string;
  state: string;
  city: string;
  address_line_1: string;
  address_line_2: string;
  postal_code: string;
  customer_group?: string;
  gst_number?: string;
}